Create ClassConcepts.cpp
A Class is a user-defined data type that has data members and member functions.
Data members are the data variables and member functions are the functions used to manipulate these variables together, these data members and member functions define the properties and behavior of the objects in a Class.

Example of class Car, the data member will be speed limit, mileage, etc, and member functions can be applying brakes, increasing speed, etc

An Object is an instance of a Class. When a class is defined, no memory is allocated but when it is instantiated (i.e. an object is created) memory is allocated.